+++ dvha/main.py
@@ -14,7 +14,7 @@ import wx
 from datetime import datetime
 from dvha.db import sql_columns
 from dvha.db.sql_to_python import QuerySQL
-from dvha.db.sql_connector import echo_sql_db
+from dvha.db.sql_connector import echo_sql_db, initialize_db
 from dvha.dialogs.main import query_dlg, UserSettings, About
 from dvha.dialogs.database import SQLSettingsDialog
 from dvha.dialogs.export import ExportCSVDialog, save_data_to_file
@@ -37,7 +37,7 @@ from dvha.paths import LOGO_PATH, DATA_DIR, ICONS
 from dvha.tools.errors import MemoryErrorDialog, PlottingMemoryError
 from dvha.tools.roi_name_manager import DatabaseROIs
 from dvha.tools.stats import StatsData, sync_variables_in_stats_data_objects
-from dvha.tools.utilities import get_study_instance_uids, scale_bitmap, is_windows, is_linux, get_window_size, \
+from dvha.tools.utilities import get_study_instance_uids, scale_bitmap, is_windows, is_linux, is_mac, get_window_size, \
     save_object_to_file, load_object_from_file, set_msw_background_color, initialize_directories_and_settings
 from dvha.db.sql_columns import all_columns as sql_column_info
 
@@ -178,7 +178,8 @@ class DVHAMainFrame(wx.Frame):
         settings_menu = wx.Menu()
         menu_pref = settings_menu.Append(wx.ID_PREFERENCES)
         menu_sql = settings_menu.Append(wx.ID_ANY, '&Database Connection\tCtrl+D')
-        menu_user_settings = settings_menu.Append(wx.ID_ANY, '&User Settings\tCtrl+,')
+        if is_mac():
+            menu_user_settings = settings_menu.Append(wx.ID_ANY, '&Preferences\tCtrl+,')
 
         help_menu = wx.Menu()
         menu_about = help_menu.Append(wx.ID_ANY, '&About')
@@ -189,9 +190,10 @@ class DVHAMainFrame(wx.Frame):
         self.Bind(wx.EVT_MENU, self.on_export, export_csv)
         self.Bind(wx.EVT_MENU, self.on_save, menu_save)
         self.Bind(wx.EVT_MENU, self.on_pref, menu_pref)
-        self.Bind(wx.EVT_MENU, self.on_pref, menu_user_settings)
         self.Bind(wx.EVT_MENU, self.on_about, menu_about)
         self.Bind(wx.EVT_MENU, self.on_sql, menu_sql)
+        if is_mac():
+            self.Bind(wx.EVT_MENU, self.on_pref, menu_user_settings)
 
         self.Bind(wx.EVT_MENU, self.on_save_plot_dvhs, export_dvhs)
         self.Bind(wx.EVT_MENU, self.on_save_plot_time_series, export_time_series)
@@ -435,14 +437,17 @@ class DVHAMainFrame(wx.Frame):
         else:
             self.button_query_execute.Disable()
 
-        # Force use to populate group 1 first
+        # Force user to populate group 1 first
         if self.radio_button_query_group.GetSelection() == 1 and self.group_data[1]['dvh'] is None:
             self.button_query_execute.Disable()
 
     def __catch_failed_sql_connection_on_app_launch(self):
-        if not echo_sql_db():
-            wx.MessageBox('Invalid credentials!', 'Echo SQL Database', wx.OK | wx.ICON_WARNING)
-            self.on_sql()
+        if self.options.DB_TYPE == 'pgsql':
+            if not echo_sql_db():
+                wx.MessageBox('Invalid credentials!', 'Echo SQL Database', wx.OK | wx.ICON_WARNING)
+                self.on_sql()
+        else:  # if using sqlite
+            initialize_db()
 
     # --------------------------------------------------------------------------------------------------------------
     # Menu bar event functions
@@ -794,7 +799,7 @@ class DVHAMainFrame(wx.Frame):
         UserSettings(self.options)
 
     def on_sql(self, *args):
-        SQLSettingsDialog()
+        SQLSettingsDialog(self.options)
         [self.__disable_add_filter_buttons, self.__enable_add_filter_buttons][echo_sql_db()]()
 
     def on_save_plot_dvhs(self, evt):