+++ dvha/options.py
@@ -14,6 +14,7 @@ import pickle
 from os.path import isfile
 from os import unlink
 import hashlib
+from copy import deepcopy
 from dvha.paths import OPTIONS_PATH, OPTIONS_CHECKSUM_PATH
 
 
@@ -22,7 +23,16 @@ class DefaultOptions:
     Create default options, to be inherited by Options class
     """
     def __init__(self):
-        self.VERSION = '0.6.6'
+        self.VERSION = '0.6.7rc2'
+
+        self.DB_TYPE = 'sqlite'
+        self.SQL_PGSQL_IP_HIST = []
+        self.DEFAULT_CNF = {'pgsql': {'host': 'localhost',
+                                      'dbname': 'dvh',
+                                      'port': '5432'},
+                            'sqlite': {'host': 'dvha.db'}}
+        self.SQL_LAST_CNX = deepcopy(self.DEFAULT_CNF)
+        self._sql_vars = ['DB_TYPE', 'SQL_PGSQL_IP_HIST', 'DEFAULT_CNF', 'SQL_LAST_CNX']
 
         self.MIN_BORDER = 50
 
@@ -180,6 +190,7 @@ class Options(DefaultOptions):
         out_options = {}
         for attr in self.option_attr:
             out_options[attr] = getattr(self, attr)
+        out_options['VERSION'] = DefaultOptions().VERSION
         with open(OPTIONS_PATH, 'wb') as outfile:
             pickle.dump(out_options, outfile)
         self.save_checksum()
@@ -215,7 +226,7 @@ class Options(DefaultOptions):
             stored_checksum = self.load_stored_checksum()
             if current_checksum == stored_checksum:
                 return True
-        except:
+        except Exception:
             pass
         print('Corrupted options file detected. Loading default options.')
         return False
@@ -228,5 +239,6 @@ class Options(DefaultOptions):
         default_options = DefaultOptions()
 
         for attr in default_options.__dict__:
-            if not attr.startswith('_'):
+            if not attr.startswith('_') and attr not in self._sql_vars:
                 setattr(self, attr, getattr(default_options, attr))
+